+++ src/@lib/equalities/shallowEquals.ts
@@ -1,3 +1,30 @@
+// shallowEquals 함수는 두 값의 얕은 비교를 수행합니다.
+
 export function shallowEquals<T>(objA: T, objB: T): boolean {
-  return objA === objB;
+  // 1. 두 값이 정확히 같은지 확인 (참조가 같은 경우)
+  if (objA === objB) {
+    return true;
+  }
+
+  // 2. 둘 중 하나라도 객체가 아닌 경우 처리
+  if (typeof objA !== "object" || typeof objB !== "object") {
+    return false;
+  }
+
+  // 3. 객체의 키 개수가 다른 경우 처리
+  const keysA = Object.keys(objA);
+  const keysB = Object.keys(objB);
+  
+  if (keysA.length !== keysB.length) {
+    return false;
+  }
+
+  // 4. 모든 키에 대해 얕은 비교 수행
+  for (const key of keysA) {
+    if (!Object.prototype.hasOwnProperty.call(objB, key) || objA[key] !== objB[key]) {
+      return false;
+    }
+  }
+  // 이 부분을 적절히 수정하세요.
+  return true;
 }
