+++ src/@lib/hooks/useCallback.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars,@typescript-eslint/no-unsafe-function-type */
+import { useMemo } from "./useMemo";
 import { DependencyList } from "react";
 
 export function useCallback<T extends Function>(
@@ -6,5 +7,5 @@ export function useCallback<T extends Function>(
   _deps: DependencyList,
 ) {
   // 직접 작성한 useMemo를 통해서 만들어보세요.
-  return factory as T;
+  return useMemo(() => factory, _deps);
 }
